Golden Rabbit Critter
This mod introduces the Golden Rabbit critter that can spawn in the same biome as the regular Rabbit, but it behaves as the Leprechaun (as in it runs aways when it sees the player). It drops a few coins as a suggestion that it will reward the player if you _*ahem*_ -- capture it.